separate StreamFilteringTestBase to its own class
