+++ test/Tester/StreamingTests/StreamFilteringTestsBase.cs
@@ -2,16 +2,13 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
-using FluentAssertions;
 using Orleans;
 using Orleans.Runtime;
-using Orleans.Runtime.Configuration;
-using Orleans.TestingHost;
 using Orleans.TestingHost.Utils;
 using TestExtensions;
 using UnitTests.GrainInterfaces;
 using UnitTests.StreamingTests;
-using Xunit;
+using FluentAssertions;
 
 namespace Tester.StreamingTests
 {
@@ -224,61 +221,6 @@ namespace Tester.StreamingTests
             count = await consumer.GetReceivedCount();
             logger.Info("Received count = {0} after third send for consumer {1}", count, consumer);
             count.Should().Be(expectedCount, "Expected count after second send");
-        }
-    }
-
-    public class StreamFilteringTests_SMS : StreamFilteringTestsBase, IClassFixture<StreamFilteringTests_SMS.Fixture>
-    {
-        public class Fixture : BaseTestClusterFixture
-        {
-            public const string StreamProvider = StreamTestsConstants.SMS_STREAM_PROVIDER_NAME;
-            protected override TestCluster CreateTestCluster()
-            {
-                var options = new TestClusterOptions(2);
-
-                options.ClusterConfiguration.AddMemoryStorageProvider("MemoryStore");
-                options.ClusterConfiguration.AddMemoryStorageProvider("PubSubStore");
-
-                options.ClusterConfiguration.AddSimpleMessageStreamProvider(StreamProvider, false);
-                options.ClientConfiguration.AddSimpleMessageStreamProvider(StreamProvider, false);
-                return new TestCluster(options);
-            }
-        }
-
-        public StreamFilteringTests_SMS()
-        {
-            streamProviderName = Fixture.StreamProvider;
-        }
-
-        [Fact, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Streaming"), TestCategory("Filters")]
-        public async Task SMS_Filter_Basic()
-        {
-            await Test_Filter_EvenOdd(true);
-        }
-
-        [Fact, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Streaming"), TestCategory("Filters")]
-        public async Task SMS_Filter_EvenOdd()
-        {
-            await Test_Filter_EvenOdd();
-        }
-
-        [Fact, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Streaming"), TestCategory("Filters")]
-        public async Task SMS_Filter_BadFunc()
-        {
-            await Assert.ThrowsAsync(typeof(ArgumentException), () =>
-                 Test_Filter_BadFunc());
-        }
-
-        [Fact, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Streaming"), TestCategory("Filters")]
-        public async Task SMS_Filter_TwoObsv_Different()
-        {
-            await Test_Filter_TwoObsv_Different();
-        }
-
-        [Fact, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Streaming"), TestCategory("Filters")]
-        public async Task SMS_Filter_TwoObsv_Same()
-        {
-            await Test_Filter_TwoObsv_Same();
         }
     }
 }
